Delete file 'WhenCreatingCourseViewModel'
- Add the two tests to different, more pertinent files in the CoursesViewModelTest folder
